[Mixpanel]:: Refactored code.
