+++ v0/destinations/mp/transform.js
@@ -1,12 +1,10 @@
 const get = require("get-value");
-const set = require("set-value");
 const { EventType } = require("../../../constants");
 const {
   removeUndefinedValues,
   defaultRequestConfig,
   defaultPostRequestConfig,
   constructPayload,
-  isDefined,
   getBrowserInfo,
   getValuesAsArrayFromConfig,
   toUnixTimestamp,
@@ -14,17 +12,9 @@ const {
   getErrorRespEvents,
   getSuccessRespEvents,
   CustomError,
-  isAppleFamily,
-  getFullName,
-  extractCustomFields,
   isHttpStatusSuccess
 } = require("../../util");
-const {
-  ConfigCategory,
-  mappingConfig,
-  MP_IDENTIFY_EXCLUSION_LIST,
-  GEO_SOURCE_ALLOWED_VALUES
-} = require("./config");
+const { ConfigCategory, mappingConfig } = require("./config");
 
 const { httpPOST } = require("../../../adapters/network");
 
@@ -32,11 +22,9 @@ const {
   processAxiosResponse
 } = require("../../../adapters/utils/networkUtils");
 
+const { createIdentifyResponse } = require("./util");
+
 // ref: https://help.mixpanel.com/hc/en-us/articles/115004613766-Default-Properties-Collected-by-Mixpanel
-const mPIdentifyConfigJson = mappingConfig[ConfigCategory.IDENTIFY.name];
-const mPProfileAndroidConfigJson =
-  mappingConfig[ConfigCategory.PROFILE_ANDROID.name];
-const mPProfileIosConfigJson = mappingConfig[ConfigCategory.PROFILE_IOS.name];
 const mPEventPropertiesConfigJson =
   mappingConfig[ConfigCategory.EVENT_PROPERTIES.name];
 
@@ -209,94 +197,21 @@ function processTrack(message, destination) {
   return returnValue;
 }
 
-function getTransformedJSON(message, mappingJson, useNewMapping) {
-  let rawPayload = constructPayload(message, mappingJson);
-  if (
-    isDefined(rawPayload.$geo_source) &&
-    !GEO_SOURCE_ALLOWED_VALUES.includes(rawPayload.$geo_source)
-  ) {
-    throw new CustomError(
-      "$geo_source value must be either null or 'reverse_geocoding' ",
-      400
-    );
-  }
-  const userName = get(rawPayload, "$name");
-  if (!userName) {
-    set(rawPayload, "$name", getFullName(message));
-  }
-
-  rawPayload = extractCustomFields(
-    message,
-    rawPayload,
-    ["traits", "context.traits"],
-    MP_IDENTIFY_EXCLUSION_LIST
-  );
-  /*
-  we are adding backward compatibility using useNewMapping key.
-  TODO :: This portion need to be removed after we deciding to stop 
-  support for old mapping.
-  */
-
-  if (!useNewMapping) {
-    if (rawPayload.$first_name) {
-      rawPayload.$firstName = rawPayload.$first_name;
-      delete rawPayload.$first_name;
-    }
-    if (rawPayload.$last_name) {
-      rawPayload.$lastName = rawPayload.$last_name;
-      delete rawPayload.$last_name;
-    }
-  }
-
-  return rawPayload;
-}
-
 async function processIdentifyEvents(message, type, destination) {
   let returnValue;
-  // this variable is used for supporting backward compatibility
-  const { useNewMapping } = destination.Config;
-  // user payload created
-  let properties = getTransformedJSON(
-    message,
-    mPIdentifyConfigJson,
-    useNewMapping
-  );
-  const device = get(message, "context.device");
-  if (device && device.token) {
-    let payload;
-    if (isAppleFamily(device.type)) {
-      payload = constructPayload(message, mPProfileIosConfigJson);
-      properties.$ios_devices = [device.token];
-    } else if (device.type.toLowerCase() === "android") {
-      payload = constructPayload(message, mPProfileAndroidConfigJson);
-      properties.$android_devices = [device.token];
-    }
-    properties = { ...properties, ...payload };
-  }
-  if (message.channel === "web" && message.context?.userAgent) {
-    const browser = getBrowserInfo(message.context.userAgent);
-    properties.$browser = browser.name;
-    properties.$browser_version = browser.version;
-  }
-  const unixTimestamp = toUnixTimestamp(message.timestamp);
 
-  const parameters = {
-    $set: properties,
-    $token: destination.Config.token,
-    $distinct_id: message.userId || message.anonymousId,
-    $ip: get(message, "context.ip") || message.request_ip,
-    $time: unixTimestamp
-  };
-  if (message.context?.active === false) {
-    parameters.$ignore_time = true;
-  }
-  returnValue = responseBuilderSimple(
-    parameters,
+  // Creating the response to identify an user
+  // https://developer.mixpanel.com/reference/profile-set
+  returnValue = createIdentifyResponse(
     message,
     type,
-    destination.Config
+    destination,
+    responseBuilderSimple
   );
 
+  // If userId and anonymousId both are present and required credentials for /import
+  // endpoint are available we are creating the merging response below
+  // https://developer.mixpanel.com/reference/identity-merge
   if (
     message.userId &&
     message.anonymousId &&
@@ -333,14 +248,13 @@ function processPageOrScreenEvents(message, type, destination) {
     message,
     mPEventPropertiesConfigJson
   );
-  const unixTimestamp = toUnixTimestamp(message.timestamp);
   const properties = {
     ...get(message, "context.traits"),
     ...message.properties,
     ...mappedProperties,
     token: destination.Config.token,
     distinct_id: message.userId || message.anonymousId,
-    time: unixTimestamp
+    time: toUnixTimestamp(message.timestamp)
   };
 
   if (message.name) {
